requirements.txt
GitPython
coloredlogs
numpy
SoundFile
scipy
voicefixer
scikit-image

librosa
torchlibrosa==0.0.7
matplotlib
setuptools
setuptools-scm
tensorboardX
pillow
julius
diffq

soundfile
typing
pynvml
progressbar
